backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=1.1.0
tiktoken>=0.5.0
# Optional dependencies - install with: pip install python-magic
# python-magic>=0.4.27
# unstructured[pdf,docx]>=0.10.0
# nltk>=3.8